+++ sharding-core/sharding-core-optimize/src/main/java/org/apache/shardingsphere/core/optimize/segment/table/TablesContext.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.core.optimize.segment;
+package org.apache.shardingsphere.core.optimize.segment.table;
 
 import com.google.common.base.Optional;
 import com.google.common.base.Preconditions;
