+++ sharding-core/sharding-core-optimize/src/test/java/org/apache/shardingsphere/core/optimize/segment/table/TablesTest.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.core.optimize.segment;
+package org.apache.shardingsphere.core.optimize.segment.table;
 
 import com.google.common.base.Optional;
 import com.google.common.collect.Sets;
